Refactor: breakup DFA.hs with LLA.hs
